ml/requirements.txt
ultralytics
opencv-python
groq
python-dotenv
